# spec/openapi.json
{
  "components": {
    "schemas": {
      "InterestsRequest": {
        "description": "Interest theme update payload for `PUT /api/v1/users/me/interests`.",
        "properties": {
          "interestThemeIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "type": "array"
          }
        },
        "required": [
          "interestThemeIds"
        ],
        "type": "object"
      },
      "LoginRequest": {
        "description": "Login request body for `POST /api/v1/login`.\n\nExample JSON:\n`{\"username\":\"admin\",\"password\":\"password\"}`",
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
      "crate.domain.Error": {
        "description": "OpenAPI schema for [`crate::domain::Error`].\n\nAPI error response payload with machine-readable code and human-readable\nmessage.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/crate.domain.ErrorCode",
            "description": "Stable machine-readable error code."
          },
          "details": {
            "description": "Supplementary error details for clients."
          },
          "message": {
            "description": "Human-readable message returned to clients.",
            "example": "Something went wrong",
            "type": "string"
          },
          "trace_id": {
            "description": "Correlation identifier for tracing this error across systems.",
            "example": "01HZY8B2W6X5Y7Z9ABCD1234",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "crate.domain.ErrorCode": {
        "description": "OpenAPI schema for [`crate::domain::ErrorCode`].\n\nStable machine-readable error codes returned in API error responses.",
        "enum": [
          "invalid_request",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "service_unavailable",
          "internal_error"
        ],
        "type": "string"
      },
      "crate.domain.InterestThemeId": {
        "description": "OpenAPI schema for [`crate::domain::InterestThemeId`].\n\nInterest theme identifiers are UUIDs serialized as strings.",
        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
        "format": "uuid",
        "type": "string"
      },
      "crate.domain.User": {
        "description": "OpenAPI schema for [`crate::domain::User`].\n\nApplication user with stable identifier and display name.",
        "properties": {
          "display_name": {
            "description": "Display name shown to other users.\n\nSchema constraints: 3–32 characters, alphanumeric plus spaces and\nunderscores. The domain layer additionally validates that the value\nis non-empty when trimmed.",
            "example": "Ada Lovelace",
            "maxLength": 32,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9_ ]+$",
            "type": "string"
          },
          "id": {
            "description": "Stable user identifier.\n\nMatches the domain `User` invariant of being a UUID.",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "type": "object"
      },
      "crate.domain.UserInterests": {
        "description": "OpenAPI schema for [`crate::domain::UserInterests`].\n\nUser interest selections with theme identifiers.",
        "properties": {
          "interestThemeIds": {
            "description": "Selected interest theme identifiers.",
            "items": {
              "$ref": "#/components/schemas/crate.domain.InterestThemeId"
            },
            "maxItems": 100,
            "type": "array"
          },
          "user_id": {
            "description": "Stable user identifier.",
            "example": "11111111-1111-1111-1111-111111111111",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "interestThemeIds"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie issued by POST /api/v1/login.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "HTTP interface for session-authenticated access and health probes.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Wildside backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/login": {
      "post": {
        "description": "Uses the centralised `Error` type so clients get a consistent\nerror schema across all endpoints.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login success",
            "headers": {
              "Set-Cookie": {
                "description": "Session cookie",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {}
        ],
        "summary": "Authenticate user and establish a session.",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "# Examples\n```\nuse actix_web::App;\nuse backend::inbound::http::users::list_users;\n\nlet app = App::new().service(list_users);\n```",
        "operationId": "listUsers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/crate.domain.User"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Users"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List known users.",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "currentUser",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.User"
                }
              }
            },
            "description": "User profile"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Fetch the authenticated user's profile.",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/interests": {
      "put": {
        "operationId": "updateUserInterests",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterestsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.UserInterests"
                }
              }
            },
            "description": "Updated interests"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Unauthorised"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.domain.Error"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Update the authenticated user's interest theme selections.",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the deployment base URL",
      "url": "/"
    }
  ],
...
}